Moved connection related handler into dedicated package
