+++ server/src/main/java/com/consol/citrus/db/server/JdbcServer.java
@@ -22,9 +22,7 @@ import com.consol.citrus.db.server.controller.RuleBasedController;
 import com.consol.citrus.db.server.controller.SimpleJdbcController;
 import com.consol.citrus.db.server.exceptionhandler.JdbcServerExceptionHandler;
 import com.consol.citrus.db.server.handler.CloseCallableStatementHandler;
-import com.consol.citrus.db.server.handler.CloseConnectionHandler;
 import com.consol.citrus.db.server.handler.CloseStatementHandler;
-import com.consol.citrus.db.server.handler.CommitTransactionStatementsHandler;
 import com.consol.citrus.db.server.handler.CreateCallableStatementHandler;
 import com.consol.citrus.db.server.handler.CreatePreparedStatementHandler;
 import com.consol.citrus.db.server.handler.CreateStatementHandler;
@@ -33,10 +31,12 @@ import com.consol.citrus.db.server.handler.ExecuteJsonQueryHandler;
 import com.consol.citrus.db.server.handler.ExecuteStatementHandler;
 import com.consol.citrus.db.server.handler.ExecuteUpdateHandler;
 import com.consol.citrus.db.server.handler.ExecuteXmlQueryHandler;
-import com.consol.citrus.db.server.handler.GetTransactionStateHandler;
-import com.consol.citrus.db.server.handler.OpenConnectionHandler;
-import com.consol.citrus.db.server.handler.RollbackTransactionStatementsHandler;
-import com.consol.citrus.db.server.handler.SetTransactionStateHandler;
+import com.consol.citrus.db.server.handler.connection.CloseConnectionHandler;
+import com.consol.citrus.db.server.handler.connection.CommitTransactionStatementsHandler;
+import com.consol.citrus.db.server.handler.connection.GetTransactionStateHandler;
+import com.consol.citrus.db.server.handler.connection.OpenConnectionHandler;
+import com.consol.citrus.db.server.handler.connection.RollbackTransactionStatementsHandler;
+import com.consol.citrus.db.server.handler.connection.SetTransactionStateHandler;
 import com.consol.citrus.db.server.transformer.JsonResponseTransformer;
 import com.consol.citrus.db.server.transformer.XmlResponseTransformer;
 import org.slf4j.Logger;
