+++ server/src/main/java/com/consol/citrus/db/server/handler/connection/CloseConnectionHandler.java
@@ -14,21 +14,23 @@
  * limitations under the License.
  */
 
-package com.consol.citrus.db.server.handler;
+package com.consol.citrus.db.server.handler.connection;
 
 import com.consol.citrus.db.server.controller.JdbcController;
+import com.consol.citrus.db.server.handler.AbstractJdbcRequestHandler;
 import spark.Request;
 import spark.Response;
 
-public class GetTransactionStateHandler extends AbstractJdbcRequestHandler {
+public class CloseConnectionHandler extends AbstractJdbcRequestHandler {
 
 
-    public GetTransactionStateHandler(final JdbcController controller) {
+    public CloseConnectionHandler(final JdbcController controller) {
         super(controller);
     }
 
     @Override
     public Object handle(final Request request, final Response response){
-        return controller.getTransactionState();
+        controller.closeConnection();
+        return "";
     }
 }
