+++ server/src/main/java/com/consol/citrus/db/server/handler/connection/CommitTransactionStatementsHandler.java
@@ -14,22 +14,23 @@
  * limitations under the License.
  */
 
-package com.consol.citrus.db.server.handler;
+package com.consol.citrus.db.server.handler.connection;
 
 import com.consol.citrus.db.server.controller.JdbcController;
+import com.consol.citrus.db.server.handler.AbstractJdbcRequestHandler;
 import spark.Request;
 import spark.Response;
 
-public class CloseConnectionHandler extends AbstractJdbcRequestHandler {
+public class CommitTransactionStatementsHandler extends AbstractJdbcRequestHandler {
 
 
-    public CloseConnectionHandler(final JdbcController controller) {
+    public CommitTransactionStatementsHandler(final JdbcController controller) {
         super(controller);
     }
 
     @Override
-    public Object handle(final Request request, final Response response){
-        controller.closeConnection();
+    public Object handle(final Request request, final Response response) {
+        controller.commitStatements();
         return "";
     }
 }
