+++ server/src/main/java/com/consol/citrus/db/server/handler/connection/GetTransactionStateHandler.java
@@ -14,22 +14,22 @@
  * limitations under the License.
  */
 
-package com.consol.citrus.db.server.handler;
+package com.consol.citrus.db.server.handler.connection;
 
 import com.consol.citrus.db.server.controller.JdbcController;
+import com.consol.citrus.db.server.handler.AbstractJdbcRequestHandler;
 import spark.Request;
 import spark.Response;
 
-public class SetTransactionStateHandler extends AbstractJdbcRequestHandler {
+public class GetTransactionStateHandler extends AbstractJdbcRequestHandler {
 
 
-    public SetTransactionStateHandler(final JdbcController controller) {
+    public GetTransactionStateHandler(final JdbcController controller) {
         super(controller);
     }
 
     @Override
-    public Object handle(final Request request, final Response response) {
-        controller.setTransactionState(Boolean.valueOf(request.body()));
-        return "";
+    public Object handle(final Request request, final Response response){
+        return controller.getTransactionState();
     }
 }
