+++ server/src/test/java/com/consol/citrus/db/server/handler/connection/CommitTransactionStatementsHandlerTest.java
@@ -14,44 +14,33 @@
  * limitations under the License.
  */
 
-package com.consol.citrus.db.server.handler;
+package com.consol.citrus.db.server.handler.connection;
 
 import com.consol.citrus.db.server.controller.JdbcController;
 import org.testng.annotations.Test;
 import spark.Request;
 import spark.Response;
 
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
-public class OpenConnectionHandlerTest {
+public class CommitTransactionStatementsHandlerTest {
 
     private final JdbcController controllerMock = mock(JdbcController.class);
-    private final OpenConnectionHandler openConnectionHandler = new OpenConnectionHandler(controllerMock);
+    private final CommitTransactionStatementsHandler commitTransactionStatementsHandler =
+            new CommitTransactionStatementsHandler(controllerMock);
 
     @Test
     public void testControllerIsUsed(){
 
         //GIVEN
         final Request requestMock = mock(Request.class);
-        when(requestMock.queryParams()).thenReturn(Collections.singleton("someKey"));
-        when(requestMock.queryParams("someKey")).thenReturn("someValue");
-
-        final Map<String, String> expectedMap = new HashMap<>();
-        expectedMap.put("someKey", "someValue");
-
         final Response responseMock = mock(Response.class);
 
         //WHEN
-        openConnectionHandler.handle(requestMock, responseMock);
+        commitTransactionStatementsHandler.handle(requestMock, responseMock);
 
         //THEN
-        verify(controllerMock).openConnection(expectedMap);
+        verify(controllerMock).commitStatements();
     }
-
 }
